+++ ui/lib/client/translations/en.yaml
@@ -19,6 +19,7 @@ error:
       ng_monitoring_not_start: |
         To use or learn more about "Continuous Profiling" feature, please search for "Continuous Profiling" in the {{distro.tidb}} official docs for more information.
         If it doesn't resove the issue, please contact the product's technical support.
+    feature_not_supported: The cluster of this version doesn't support or can't use this feature, please contact with technical support to get more information.
     other: Other error
   tidb:
     no_alive_tidb: No alive {{distro.tidb}} instance
